+++ src/main/java/insane96mcp/enhancedai/modules/zombie/ai/DiggingGoal.java
@@ -163,6 +163,9 @@ public class DiggingGoal extends Goal {
 			if (DiggerZombie.blockBlacklist.isBlockBlackOrNotWhiteListed(state.getBlock()))
 				continue;
 
+			if (state.hasBlockEntity() && DiggerZombie.blacklistTileEntities)
+				continue;
+
 			this.targetBlocks.add(rayTraceResult.getBlockPos());
 		}
 		Collections.reverse(this.targetBlocks);